STM32H7 directory cleanup and refactoring
