+++ targets/TARGET_STM/TARGET_STM32H7/TARGET_STM32H743xI/flash_data.h
