+++ targets/TARGET_STM/TARGET_STM32H7/TARGET_STM32H743xI/us_ticker_data.h
@@ -23,7 +23,7 @@
 #include "stm32h7xx.h"
 #include "stm32h7xx_ll_tim.h"
 #include "cmsis_nvic.h"
-   
+
 #define TIM_MST      TIM5
 #define TIM_MST_IRQ  TIM5_IRQn
 #define TIM_MST_RCC  __TIM5_CLK_ENABLE()
