+++ targets/TARGET_STM/TARGET_STM32H7/device/stm32h7xx_hal_rtc.h
@@ -27,7 +27,6 @@
 
 /* Includes ------------------------------------------------------------------*/
 #include "stm32h7xx_hal_def.h"
-#include "stm32h7xx_ll_rtc.h" // MBED PATCH
 
 /** @addtogroup STM32H7xx_HAL_Driver
   * @{
